requirements.txt
kelvin-python-sdk
structlog